Move SetBonusDisplay into SetBonus.tsx
